added code changes to integrate phonepe

+++ src/main/java/com/app/movie/trade/entities/DealDetailInfo.java
@@ -11,6 +11,6 @@ public interface DealDetailInfo {
 	 String getTheatrename();
 	 LocalDate getShowdate();
 	 LocalTime getShowtime();
-	 int getTotaldealprice();
+	 double getTotaldealprice();
 	 int getCapacity();
 }
